refactor(suite-native): extract useTapGesture so it's reusable
